+++ CHANGELOG.md
@@ -16,9 +16,9 @@ Types of changes
 
 ## [Unreleased]
 
-- `Changed` improve masking.yml readability and conciceness (possible solutions: merge frequent mask combination into single use e.g. add with template=>add with template syntax support ; include external yaml to deduplcate, parameterized masks, ...).
+- Nothing in the backlog :)
 
-## [1.3.0 (Unreleased)]
+## [1.3.0]
 
 - `Added` ff1 mask to meet the requirement of re-identification from a single secret key.
 - `Fixed` use same random mask on different path produce the same value.